Refactor animation structures by separating Color and Transform into their own modules
